test: refactored skill seeder for auomated tests for skills
